Updated datatype for Frequency and PaymentMode SPs

+++ Development/Source Code/Sparkle/Sparkle/Database/Configuration/Stored Procedures/FrequencyDelete.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROCEDURE [Configuration].[FrequencyDelete]
 (
-	@Id Numeric(10,0)
+	@Id BIGINT
 )
 AS
 BEGIN
